Midified: added user roles

+++ Services/AuthorService/AuthorService.cs
@@ -1,12 +1,14 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using Book_Store.Data;
 using Book_Store.Dtos.Product_Entities.Author;
 using Book_Store.Models;
 using Book_Store.Models.Product_Entities;
+using Book_Store.Models.User;
 using Microsoft.EntityFrameworkCore;
 
 namespace Book_Store.Services.AuthorService
@@ -15,12 +17,15 @@ namespace Book_Store.Services.AuthorService
     {
         private readonly IMapper mapper;
         public readonly DataContext context;
+        private readonly IHttpContextAccessor httpContextAccessor;
 
-        public AuthorService(IMapper mapper, DataContext context)
+        public AuthorService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             this.context = context;
+            this.httpContextAccessor = httpContextAccessor;
             this.mapper = mapper;
-        }
+        } 
+        private int GetUserId() => Convert.ToInt32(httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetAuthorDto>>> GetAllAuthors()
         {
@@ -33,6 +38,14 @@ namespace Book_Store.Services.AuthorService
         public async Task<ServiceResponse<List<GetAuthorDto>>> AddAuthor(AddAuthorDto newAuthor)
         {
             var serviceResponse = new ServiceResponse<List<GetAuthorDto>>();
+            User currentUser = await context.User.FirstAsync(c => c.Id == GetUserId());
+            if(currentUser.Role != Role.Seller)
+            {
+                serviceResponse.Success = false;
+                serviceResponse.Message = "Permission denied";
+                return serviceResponse;
+            }
+
             Author author = mapper.Map<Author>(newAuthor);
 
             context.Author.Add(author);
@@ -48,6 +61,14 @@ namespace Book_Store.Services.AuthorService
 
             try
             {
+                User currentUser = await context.User.FirstAsync(c => c.Id == GetUserId());
+                if(currentUser.Role != Role.Seller)
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Permission denied";
+                    return serviceResponse;
+                }
+
                 var authorToUpdate = await context.Author.FirstAsync(c => c.Id == id);
                 mapper.Map(updatedAuthor, authorToUpdate);
                 await context.SaveChangesAsync();
@@ -68,6 +89,14 @@ namespace Book_Store.Services.AuthorService
 
             try
             {
+                User currentUser = await context.User.FirstAsync(c => c.Id == GetUserId());
+                if(currentUser.Role != Role.Seller)
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Permission denied";
+                    return serviceResponse;
+                }
+
                 var author = await context.Author.FirstAsync(c => c.Id == id);
                 context.Author.Remove(author);
                 await context.SaveChangesAsync();
